Reworked structure.
Telegram wrapper separated to independed file, compatible to innightwolfsleep/llama-cpp-telegram_bot
